feat(item): add new item
