+++ lib/statistics.ts
@@ -1,7 +1,4 @@
-import { get } from 'lodash';
-
-import { GENERAL_CONSTANTS, HTTP_ERRORS } from './constants';
-import { UserDetails, UserLoginDetails, LoggedInUserDetails } from '../lib/types';
+import { UserDetails, UserLoginDetails } from '../lib/types';
 
 export async function fetchCategories(values: UserDetails): Promise<void> {
     try {
